Reset password

+++ app/User.php
@@ -59,7 +59,9 @@ class User extends Authenticatable implements MustVerifyEmail
 
         static::deleting(function($user) {
             // Delete verify token if exist
-            $user->verifyUser->delete();
+            if ($user->verifyUser){
+                $user->verifyUser->delete();
+            }
             // Delete user roles
             foreach($user->roles as $role){
                 $user->roles()->detach($role->id);
